+++ cw-orch-interchain/src/lib.rs
@@ -1,8 +1,15 @@
 // prelude
 #[cfg(not(target_arch = "wasm32"))]
 pub mod prelude {
-    pub use cw_orch_interchain_core::{IbcQueryHandler, InterchainEnv};
+    pub use cw_orch_interchain_core::{
+        types::ChannelCreationResult, IbcQueryHandler, InterchainEnv,
+    };
     pub use cw_orch_interchain_mock::{MockBech32InterchainEnv, MockInterchainEnv};
+
+    pub use ibc_relayer_types::core::{
+        ics04_channel::packet::Sequence,
+        ics24_host::identifier::{ChannelId, PortId},
+    };
 
     #[cfg(feature = "daemon")]
     pub use cw_orch_interchain_daemon::{