+++ cw-orch/tests/interface_macro.rs
@@ -95,7 +95,7 @@ fn test_migrate() {
     contract.upload().unwrap();
 
     contract
-        .instantiate(&InstantiateMsg {}, Some(&chain.sender()), None)
+        .instantiate(&InstantiateMsg {}, Some(&chain.sender_addr()), None)
         .unwrap();
 
     contract