+++ packages/cw-orch-core/src/environment/cosmwasm_environment.rs
@@ -26,7 +26,19 @@ pub trait TxHandler: ChainState + Clone {
     type Sender: Clone;
 
     /// Gets the address of the current wallet used to sign transactions.
+    #[deprecated(
+        since = "1.1.2",
+        note = "Please use `sender_addr` instead. This method will be changed in the next release."
+    )]
+    // TODO: return &Self::Sender here in the next breaking release
     fn sender(&self) -> Addr;
+
+    /// Gets the address of the current wallet used to sign transactions.
+    // TODO: remove this default implementation in the next breaking release
+    fn sender_addr(&self) -> Addr {
+        #[allow(deprecated)]
+        self.sender()
+    }
 
     /// Sets wallet to sign transactions.
     fn set_sender(&mut self, sender: Self::Sender);
@@ -137,6 +149,10 @@ mod tests {
         type Sender = ();
 
         fn sender(&self) -> Addr {
+            unimplemented!()
+        }
+
+        fn sender_addr(&self) -> Addr {
             unimplemented!()
         }