+++ packages/cw-orch-core/src/environment/mod.rs
@@ -13,7 +13,13 @@ pub use queriers::{
     bank::BankQuerier,
     env::{EnvironmentInfo, EnvironmentQuerier},
     node::NodeQuerier,
-    wasm::WasmQuerier,
+    wasm::{AsyncWasmQuerier, WasmQuerier},
     DefaultQueriers, Querier, QuerierGetter, QueryHandler,
 };
 pub use state::{ChainState, StateInterface};
+
+/// Describes a structure that contains an underlying execution environment
+pub trait Environment<Chain> {
+    /// Returns the chain that this structure executes on
+    fn environment(&self) -> &Chain;
+}