+++ packages/interchain/starship/src/lib.rs
@@ -6,7 +6,7 @@ pub mod client;
 use crate::client::StarshipClient;
 use cw_orch_core::environment::{ChainInfoOwned, ChainState, NetworkInfoOwned};
 use cw_orch_core::CwEnvError;
-use cw_orch_daemon::{Daemon, DaemonBuilder};
+use cw_orch_daemon::{Daemon, DaemonBuilder, RUNTIME};
 use ibc_chain_registry::chain::ChainData;
 use std::collections::HashMap;
 use tokio::runtime::Handle;
@@ -23,7 +23,13 @@ pub struct Starship {
 
 impl Starship {
     /// Creates a new instance and connects to a starship deployment
-    pub fn new(rt_handle: &Handle, url: Option<&str>) -> Result<Self, CwEnvError> {
+    pub fn new(url: Option<&str>) -> Result<Self, CwEnvError> {
+        let runtime = RUNTIME.handle();
+        Self::new_with_runtime(runtime, url)
+    }
+
+    /// Creates a new instance and connects to a starship deployment
+    pub fn new_with_runtime(rt_handle: &Handle, url: Option<&str>) -> Result<Self, CwEnvError> {
         let starship_client = StarshipClient::new(rt_handle.clone(), url)?;
 
         let mut daemons: HashMap<String, Daemon> = HashMap::new();
@@ -36,11 +42,8 @@ impl Starship {
                     .unwrap()
             });
 
-            let mut daemon_builder = DaemonBuilder::default();
-            let mut daemon_builder = daemon_builder
-                .chain(chain_data_conversion(chain.clone()))
-                .mnemonic(mnemonic)
-                .handle(rt_handle);
+            let mut daemon_builder = DaemonBuilder::new(chain_data_conversion(chain.clone()));
+            let mut daemon_builder = daemon_builder.mnemonic(mnemonic).handle(rt_handle);
 
             if let Some(existing_daemon) = daemons.values().next() {
                 daemon_builder = daemon_builder.state(existing_daemon.state())